+++ source/Desktop/DesktopTimer.csproj
@@ -98,20 +98,20 @@
   <ItemGroup>
     <Compile Include="Data\Constants.cs" />
     <Compile Include="Program.cs" />
-    <Compile Include="UiNotify.cs" />
-    <Compile Include="Pomodoro.cs">
+    <Compile Include="Views\UiNotify.cs" />
+    <Compile Include="Views\Pomodoro.cs">
       <SubType>Form</SubType>
     </Compile>
     <Compile Include="Properties\AssemblyInfo.cs" />
-    <Compile Include="TextIcon.cs" />
-    <Compile Include="TimerState.cs" />
-    <Compile Include="UpdaterForm.cs">
+    <Compile Include="Views\TextIcon.cs" />
+    <Compile Include="Views\TimerState.cs" />
+    <Compile Include="Views\UpdaterForm.cs">
       <SubType>Form</SubType>
     </Compile>
-    <Compile Include="UpdaterForm.Designer.cs">
+    <Compile Include="Views\UpdaterForm.Designer.cs">
       <DependentUpon>UpdaterForm.cs</DependentUpon>
     </Compile>
-    <EmbeddedResource Include="Pomodoro.resx">
+    <EmbeddedResource Include="Views\Pomodoro.resx">
       <DependentUpon>Pomodoro.cs</DependentUpon>
     </EmbeddedResource>
     <EmbeddedResource Include="Properties\Resources.resx">
@@ -123,7 +123,7 @@
       <AutoGen>True</AutoGen>
       <DependentUpon>Resources.resx</DependentUpon>
     </Compile>
-    <EmbeddedResource Include="UpdaterForm.resx">
+    <EmbeddedResource Include="Views\UpdaterForm.resx">
       <DependentUpon>UpdaterForm.cs</DependentUpon>
     </EmbeddedResource>
     <None Include=".licenseheader" />
