+++ source/Desktop/Views/TimerState.cs
